// samples/react-follow-document/src/webparts/followDocumentWebPart/models/followDocument.ts
export interface FollowDocument {
    id?: string;
    Title?: string;
    WebName?: string;
    WebUrl?: string;
    Domain?: string;
    Folder?: string;
    ItemProperties?: string;
    WebFileUrl?: string;
    DriveId?: string;
    ListId?: string;
    SiteId?: string;
    ItemId?: string;
    Url?: string;
    DownloadFile?: string;
    Thumbnail?: string;
    IconUrl?: string;
    documentCardActions?: Array<any>;
    preview?: string;
    Description?: string;
    followedDateTime?: Date;
}